requirements.txt
streamlit
transformers
torch
pillow
pytesseract
pandas
numpy